refactor a few util methods and update tests
