+++ server/api/experiments/[slug].get.ts
@@ -1,4 +1,4 @@
-import { getSlugOrIdPrismaWhereClause } from "~~/server/utils/utils"
+import { getSlugOrIdPrismaWhereClause } from "~~/server/utils/prisma"
 import { experimentAbilities } from "~~/shared/utils/abilities"
 import { authorize } from "~~/server/utils/authorization"
