+++ server/api/experiments/[slug].put.ts
@@ -1,5 +1,5 @@
 import type { Prisma } from "@prisma/client"
-import { getSlugOrIdPrismaWhereClause, untilSlugUnique } from "~~/server/utils/utils"
+import { getSlugOrIdPrismaWhereClause, untilSlugUnique } from "~~/server/utils/prisma"
 import slugify from "~~/server/utils/slugify"
 import { experimentAbilities } from "~~/shared/utils/abilities"
 import { authorize } from "~~/server/utils/authorization"
