+++ server/api/experiments/attributes/[slug].get.ts
@@ -1,4 +1,4 @@
-import { getSlugOrIdPrismaWhereClause } from "~~/server/utils/utils"
+import { getSlugOrIdPrismaWhereClause } from "~~/server/utils/prisma"
 import type { ExperimentAttributeDetail } from "~~/shared/types"
 import { experimentAttributeAbilities } from "~~/shared/utils/abilities"
 import { authorize } from "~~/server/utils/authorization"
