+++ server/utils/auth.ts
@@ -1,4 +1,4 @@
-import type { H3Event, EventHandlerRequest } from "h3"
+import type { Event } from "./utils"
 import { betterAuth } from "better-auth"
 import { APIError } from "better-auth/api"
 import { admin } from "better-auth/plugins"
@@ -98,7 +98,7 @@ export const auth = betterAuth({
 /**
  * getUserDetail()
  */
-export async function getUser(event: H3Event<EventHandlerRequest>): Promise<UserDetail | null> {
+export async function getUser(event: Event): Promise<UserDetail | null> {
   const session = await auth.api.getSession({
     headers: event.headers,
   })
@@ -108,7 +108,7 @@ export async function getUser(event: H3Event<EventHandlerRequest>): Promise<User
 /**
  * getUserDetailOrThrowError()
  */
-export async function getUserOrThrowError(event: H3Event<EventHandlerRequest>): Promise<UserDetail> {
+export async function getUserOrThrowError(event: Event): Promise<UserDetail> {
   const user = await getUser(event)
   if (!user) {
     throw createError({ statusCode: 401, statusMessage: "Not logged in" })
